# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BNM - Banque Nationale de Mauritanie API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Actualite {
  id: number;
  titre: string;
  slug: string;
  contenu: string;
  image?: string | null;
  categorie?: string | null;
  datePublication: string;
  createdAt: string;
}

export interface CreateActualiteInput {
  /** @maxLength 200 */
  titre: string;
  /** @maxLength 200 */
  slug: string;
  contenu: string;
  image?: string | null;
  /** @maxLength 100 */
  categorie?: string | null;
  datePublication?: string | null;
}

export type OffreCategorie =
  (typeof OffreCategorie)[keyof typeof OffreCategorie];

export const OffreCategorie = {
  particuliers: "particuliers",
  professionnels: "professionnels",
  entreprises: "entreprises",
  islamique: "islamique",
} as const;

export interface Offre {
  id: number;
  titre: string;
  slug: string;
  description?: string | null;
  icone?: string | null;
  categorie: OffreCategorie;
  ordre: number;
}

export interface ContactInput {
  /** @minLength 2 */
  nom: string;
  email: string;
  telephone?: string | null;
  sujet?: string | null;
  /** @minLength 10 */
  message: string;
}

export interface NewsletterInput {
  email: string;
}

export interface SimulateClassicInput {
  /**
   * @minimum 100000
   * @maximum 50000000
   */
  montant: number;
  /**
   * @minimum 1
   * @maximum 25
   */
  dureeAns: number;
  /**
   * @minimum 0
   * @maximum 100
   */
  taux?: number;
}

export interface SimulateMurabahaInput {
  /** @minimum 0 */
  prixBien: number;
  /** @minimum 0 */
  apportPersonnel: number;
  /**
   * @minimum 3
   * @maximum 8
   */
  marge?: number;
  /**
   * @minimum 1
   * @maximum 25
   */
  dureeAns: number;
}

export interface SimulationResult {
  mensualite: number;
  coutTotal: number;
  interetsTotal: number;
  montant: number;
  dureeAns: number;
  taux: number;
}

export interface MurabahaResult {
  mensualite: number;
  montantFinance: number;
  coutTotal: number;
  marge: number;
  dureeAns: number;
}

export interface Agence {
  id: number;
  nom: string;
  adresse?: string | null;
  ville?: string | null;
  telephone?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  horaires?: string | null;
}

export interface Stats {
  totalClients: number;
  totalAgences: number;
  anneesExperience: number;
  totalActualites: number;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface UpdateActualiteInput {
  titre?: string;
  slug?: string;
  contenu?: string;
  image?: string | null;
  categorie?: string | null;
  datePublication?: string | null;
}

export type ListActualitesParams = {
  categorie?: string;
  limit?: number;
  offset?: number;
};

export type ListActualites200 = {
  data: Actualite[];
  total: number;
};

export type ListOffresParams = {
  categorie?: ListOffresCategorie;
};

export type ListOffresCategorie =
  (typeof ListOffresCategorie)[keyof typeof ListOffresCategorie];

export const ListOffresCategorie = {
  particuliers: "particuliers",
  professionnels: "professionnels",
  entreprises: "entreprises",
  islamique: "islamique",
} as const;